--- api/src/db/seeds/development/2023.09.29T16.52.55.fill-funding-submission-lines-table.ts
@@ -1,30 +1,0 @@
-import type { SeedMigration } from "@/db/umzug"
-
-import fundingSubmissionLineData from "@/db/data/funding-submission-lines.json"
-
-export const up: SeedMigration = async ({ context: { FundingSubmissionLine } }) => {
-  const promises = fundingSubmissionLineData.map((row) => {
-    return FundingSubmissionLine.findOrCreate({
-      where: {
-        fiscalYear: row.fiscalYear,
-        sectionName: row.sectionName,
-        lineName: row.lineName,
-      },
-      defaults: {
-        fiscalYear: row.fiscalYear,
-        sectionName: row.sectionName,
-        lineName: row.lineName,
-        fromAge: row.fromAge,
-        toAge: row.toAge,
-        monthlyAmount: row.monthlyAmount,
-      },
-    })
-  })
-
-  return Promise.all(promises)
-}
-
-export const down: SeedMigration = async () => {
-  // this method needs to exist, but does not need to be implemented.
-  // Seeds should be idempotent.
-}
